requirements.txt
django == 4.2.5
djangorestframework
pyyaml
uritemplate
coreapi
requests
django-cors-headers
whitenoise == 6.5.0
pillow
gunicorn